Initializing Metrics and HTTP Monitoring Modules from separate classes
