+++ samples/metrics/src/main/java/org/wso2/carbon/mss/example/Application.java
@@ -1,21 +1,18 @@
 /*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *   WSO2 Inc. licenses this file to you under the Apache License,
- *   Version 2.0 (the "License"); you may not use this file except
- *   in compliance with the License.
- *   You may obtain a copy of the License at
- *
+ * Copyright 2015 WSO2 Inc. (http://wso2.org)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-
 package org.wso2.carbon.mss.example;
 
 import org.slf4j.Logger;
@@ -24,7 +21,6 @@ import org.wso2.carbon.mss.MicroservicesRunner;
 import org.wso2.carbon.mss.example.service.StudentService;
 import org.wso2.carbon.mss.example.service.TestService;
 import org.wso2.carbon.mss.httpmonitoring.HTTPMonitoringInterceptor;
-import org.wso2.carbon.mss.metrics.MetricReporter;
 import org.wso2.carbon.mss.metrics.MetricsInterceptor;
 
 /**
@@ -36,8 +32,8 @@ public class Application {
 
     public static void main(String[] args) {
         logger.info("Starting the Microservice with Metrics");
-        new MicroservicesRunner().addInterceptor(new HTTPMonitoringInterceptor.Builder().build())
-                .addInterceptor(new MetricsInterceptor(MetricReporter.CONSOLE, MetricReporter.JMX))
-                .deploy(new StudentService()).deploy(new TestService()).start();
+        new MicroservicesRunner().addInterceptor(new HTTPMonitoringInterceptor())
+                .addInterceptor(new MetricsInterceptor()).deploy(new StudentService()).deploy(new TestService())
+                .start();
     }
 }
