+++ samples/metrics/src/main/java/org/wso2/carbon/mss/example/service/StudentService.java
@@ -1,32 +1,34 @@
 /*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *   WSO2 Inc. licenses this file to you under the Apache License,
- *   Version 2.0 (the "License"); you may not use this file except
- *   in compliance with the License.
- *   You may obtain a copy of the License at
- *
+ * Copyright 2015 WSO2 Inc. (http://wso2.org)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-
 package org.wso2.carbon.mss.example.service;
 
 import org.wso2.carbon.metrics.annotation.Counted;
 import org.wso2.carbon.metrics.annotation.Metered;
 import org.wso2.carbon.metrics.annotation.Timed;
 import org.wso2.carbon.mss.httpmonitoring.HTTPMonitoring;
+import org.wso2.carbon.mss.httpmonitoring.HTTPMonitoringModule;
+import org.wso2.carbon.mss.metrics.MetricReporter;
+import org.wso2.carbon.mss.metrics.MetricsModule;
 
 import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -49,6 +51,18 @@ public class StudentService {
         student.setLastName("Rodgers");
         student.setAge(14);
         addStudent(student);
+    }
+
+    @PostConstruct
+    public void init() {
+        MetricsModule.init(MetricReporter.CONSOLE, MetricReporter.JMX, MetricReporter.DAS);
+        HTTPMonitoringModule.init();
+    }
+
+    @PreDestroy
+    public void destroy() {
+        MetricsModule.destroy();
+        HTTPMonitoringModule.destroy();
     }
 
     @GET
