+++ samples/metrics/src/main/java/org/wso2/carbon/mss/example/service/TestService.java
@@ -1,30 +1,32 @@
 /*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *   WSO2 Inc. licenses this file to you under the Apache License,
- *   Version 2.0 (the "License"); you may not use this file except
- *   in compliance with the License.
- *   You may obtain a copy of the License at
- *
+ * Copyright 2015 WSO2 Inc. (http://wso2.org)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-
 package org.wso2.carbon.mss.example.service;
 
 import org.wso2.carbon.metrics.annotation.Counted;
 import org.wso2.carbon.metrics.annotation.Metered;
 import org.wso2.carbon.metrics.annotation.Timed;
 import org.wso2.carbon.mss.httpmonitoring.HTTPMonitoring;
+import org.wso2.carbon.mss.httpmonitoring.HTTPMonitoringModule;
+import org.wso2.carbon.mss.metrics.MetricReporter;
+import org.wso2.carbon.mss.metrics.MetricsModule;
 
 import java.util.Random;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -39,6 +41,18 @@ public class TestService {
     private final Random random = new Random();
 
     private long total = 0L;
+
+    @PostConstruct
+    public void init() {
+        MetricsModule.init(MetricReporter.CONSOLE, MetricReporter.JMX, MetricReporter.DAS);
+        HTTPMonitoringModule.init();
+    }
+
+    @PreDestroy
+    public void destroy() {
+        MetricsModule.destroy();
+        HTTPMonitoringModule.destroy();
+    }
 
     @GET
     @Path("/rand/{bound}")
